+++ graph_tool_call/core/__init__.py
@@ -1,5 +1,5 @@
-from graph_tool_call.core.graph import NetworkXGraph
+from graph_tool_call.core.dict_graph import DictGraph
 from graph_tool_call.core.protocol import GraphEngine
 from graph_tool_call.core.tool import ToolSchema, parse_tool
 
-__all__ = ["GraphEngine", "NetworkXGraph", "ToolSchema", "parse_tool"]
+__all__ = ["DictGraph", "GraphEngine", "ToolSchema", "parse_tool"]